parse header js
